Corrected transaction bug, introduced limit and cutoff =======================================================
System Information:
Engine Version: Molecule Ver 1.2.14.20210623
Library Version:  Ver 1.0.11.20210618

Firmware:
Windows Ver 1.0.4.20210617
Linux Ver 1.0.1.20210622

Hosts:
Interactive Console Ver 1.0.4.20210622
Pipe Child Ver 1.1.1.20210622
Headless Host Ver 0.9.1.20210622

Storages:
File System Storage Ver 1.1.1.20210622

Compilers:
MSVC 14.29.30038
GCC 9.3.0

OSes:
Linux Ver:5.8.0-55-generic [x86_64]
Windows Workstation 10.0.19041 [x64]

+++ host/rx_anonymus_pipes.cpp
@@ -85,7 +85,7 @@ rx_result local_pipe_port::receive_loop (rx_pipe_host* host)
 
 
 	add_stack_endpoint(&pipes_.stack_entry_, nullptr, nullptr);
-	pipes_.receive_loop([this] (size_t count)
+	pipes_.receive_loop([this] (int64_t count)
 		{
 		});
 
@@ -99,7 +99,7 @@ rx_result local_pipe_port::start_listen (const protocol_address* local_address, 
 		RX_ASSERT(false);
 		return "Already started.";
 	}
-	auto result = pipes_.open(pipe_handles_, [this](size_t count)
+	auto result = pipes_.open(pipe_handles_, [this](int64_t count)
 		{
 		});
 	if (!result)
